+++ fronted/src/App.js
@@ -1,8 +1,7 @@
 import { useState } from "react";
 import Nav from "./components/Nav";
 import SignIn from "./components/SignIn";
-// import SignUp from "./components/SignUp";
-import SignUp from "./components/SignUp1"
+import SignUp from "./components/SignUp"
 import Home from "./components/Home";
 import HeroSection from "./components/HeroSection"
 import { BrowserRouter as Router, Route } from "react-router-dom";
@@ -23,9 +22,7 @@ function App() {
           <SignIn handleFetch={setUserData} />
         </Route>
         <Route path="/sign-up">
-          
           <SignUp handleFetch={setUserData} />
-  
         </Route>
       </Router>
     </div>
