+++ fronted/src/components/SignUp.jsx
@@ -1,12 +1,14 @@
+import logo from '../images/logo.png';
+import { Link } from 'react-router-dom';
 import { useState } from "react";
 import { useHistory } from "react-router-dom";
 
-function SignUn({ handleFetch }) {
+
+const Example = ({handleFetch}) => {
   const [user, setUser] = useState({
     email: "",
     password: "",
-    firstName: "",
-    lastName: "",
+    name: ""
   });
   let history = useHistory();
   const handlChange = (e) => {
@@ -37,54 +39,103 @@ function SignUn({ handleFetch }) {
     }
   };
   return (
-    <div>
-      <form onSubmit={handlSingUn}>
-        <div>
-          <label htmlFor="firstName">
-            First Name
-            <input
-              type="text"
-              name="firstName"
-              required
-              value={user.firstName}
-              onChange={handlChange}
-            />
-          </label>
-          <label htmlFor="firstName">
-            Last Name
-            <input
-              type="text"
-              name="lastName"
-              required
-              value={user.lastName}
-              onChange={handlChange}
-            />
-          </label>
-          <label htmlFor="email">
-            Email
-            <input
-              type="email"
-              name="email"
-              required
-              value={user.email}
-              onChange={handlChange}
-            />
-          </label>
-          <label htmlFor="password">
-            Password
-            <input
-              type="password"
-              name="password"
-              required
-              value={user.password}
-              onChange={handlChange}
-            />
-          </label>
+    <div className="min-h-screen bg-gray-50 flex flex-col justify-center py-12 sm:px-4 lg:px-6">
+      <div className="sm:mx-auto sm:w-full sm:max-w-md">
+        <img
+          className="mx-auto h-40 w-auto"
+          src={logo}
+          alt="Workflow"
+        />
+        <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">You are one step away from changing the world! <br/>Create an account and Sign up</h2>
+      </div>
+
+      <div className="mt-6 sm:mx-auto sm:w-full sm:max-w-md">
+        <div className="bg-white py-8 px-4 shadow sm:rounded-lg sm:px-10">
+          <form className="space-y-6" action="#" method="POST" onSubmit={handlSingUn}>
+          <div>
+              <label htmlFor="email" className="block text-sm font-medium text-gray-700 w-24">
+                Name
+              </label>
+              <div className="mt-1">
+                <input
+                  id="name"
+                  name="name"
+                  type="name"
+                  autoComplete="name"
+                  value={user.name}
+                  onChange={handlChange}
+                  required
+                  className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-green-500 focus:border-green-500 sm:text-sm"
+                />
+              </div>
+            </div>
+            <div>
+              <label htmlFor="email" className="block text-sm font-medium text-gray-700 w-24">
+                Email address
+              </label>
+              <div className="mt-1">
+                <input
+                  id="email"
+                  name="email"
+                  type="email"
+                  autoComplete="email"
+                  value={user.email}
+                  onChange={handlChange}
+                  required
+                  className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-green-500 focus:border-green-500 sm:text-sm"
+                />
+              </div>
+            </div>
+
+            <div>
+              <label htmlFor="password" className="block text-sm font-medium text-gray-700 w-24">
+                Password
+              </label>
+              <div className="mt-1">
+                <input
+                  id="password"
+                  name="password"
+                  type="password"
+                  autoComplete="current-password"
+                  required
+                  value={user.password}
+                  onChange={handlChange}
+                  className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-green-500 focus:border-green-500 sm:text-sm"
+                />
+              </div>
+            </div>
+
+
+            <div>
+              <button
+                type="submit"
+                className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-green-600 hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
+              >
+                Sign Up
+              </button>
+            </div>
+          </form>
+          <div className="relative my-10">
+              <div className="absolute inset-0 flex items-center">
+                <div className="w-full border-t border-gray-300" />
+              </div>
+              <div className="relative flex justify-center text-sm">
+                <span className="px-2 bg-white text-gray-500">Or If you already have an Account</span>
+              </div>
+            </div>
+            <Link to="sign-in">
+            <button
+                type="submit"
+                className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-green-400 hover:bg-green-600 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
+              >
+                Sign In
+              </button>
+              </ Link>
         </div>
-        <button type="submit">Sign Up</button>
-      </form>
+        
+      </div>
     </div>
-  );
+  )
 }
 
-export default SignUn;
+export default Example;
